属性クラス名に Attribute サフィックスを追加
