+++ src/Hinata.WebApp/Filters/AccessLogFilterAttribute.cs
@@ -1,0 +1,16 @@
+﻿using System.Web.Mvc;
+using Hinata.Logging;
+
+namespace Hinata.Filters
+{
+    public class AccessLogFilterAttribute : ActionFilterAttribute
+    {
+        private readonly IAccessLogger _webAccessLogger = LogManager.GetWebAccessLogger();
+
+        public override void OnActionExecuting(ActionExecutingContext filterContext)
+        {
+            _webAccessLogger.Write();
+            base.OnActionExecuting(filterContext);
+        }
+    }
+}
